SCRegressor: split code into selection, seed finding, filling.
